Add extraction for Tribes, including splash image
Other major changes:

* Add de-hyphenation work using a dictionary pulled from https://github.com/dwyl/english-words/blob/master/words.txt
* Tided up some other text-error detections to be more generic
* Text should now be *mostly* error free
* Added new classes for some missing location data

+++ src/main/java/name/ipsi/project/fwbp/foundry/core/DocumentOwnershipLevel.java
@@ -6,11 +6,11 @@ import java.util.HashMap;
 import java.util.Map;
 
 public enum DocumentOwnershipLevel {
-    INHERIT(-1.0),
-    NONE(0.0),
-    LIMITED(1.0),
-    OBSERVER(2.0),
-    OWNER(3.0);
+    INHERIT(-1),
+    NONE(0),
+    LIMITED(1),
+    OBSERVER(2),
+    OWNER(3);
 
     public static Map<String, DocumentOwnershipLevel> defaultInherit() {
         var value = new HashMap<String, DocumentOwnershipLevel>();
@@ -37,9 +37,9 @@ public enum DocumentOwnershipLevel {
     }
 
     @JsonValue
-    private final double jsonValue;
+    private final int jsonValue;
 
-    DocumentOwnershipLevel(double jsonValue) {
+    DocumentOwnershipLevel(int jsonValue) {
         this.jsonValue = jsonValue;
     }
 }
